Modify Flowchart
Added what to do next:    What need to make additional flowchart       Database Design for Class Check-in       Code Design for Class Check-in    What need to modify       We will require line notify to required user login, because we didn't intend to do register form for create user, and we only use In-House.    Other ideas:       In program, show error for user if they don't have permission for that field.       Redesign backend for Admin to get privilages to access other field of sites.       Teacher can Check-in using QR.       make UI easy to use for other platforms.